Revert moving reursive_desription_id
