+++ fedot/core/dag/graph_node.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
-from typing import List, Optional, Iterable, Hashable
+from typing import List, Optional, Iterable
 
-from fedot.core.dag.graph_utils import node_depth, descriptive_id_recursive
+from fedot.core.dag.graph_utils import node_depth
 
 
 class GraphNode(ABC):
@@ -75,3 +75,26 @@ class GraphNode(ABC):
             int: max depth to the primary level
         """
         return node_depth(self) - 1
+
+
+def descriptive_id_recursive(current_node: GraphNode, visited_nodes=None) -> str:
+    if visited_nodes is None:
+        visited_nodes = []
+
+    node_label = current_node.description()
+
+    full_path_items = []
+    if current_node in visited_nodes:
+        return 'ID_CYCLED'
+    visited_nodes.append(current_node)
+    if current_node.nodes_from:
+        previous_items = []
+        for parent_node in current_node.nodes_from:
+            previous_items.append(f'{descriptive_id_recursive(parent_node, copy(visited_nodes))};')
+        previous_items.sort()
+        previous_items_str = ';'.join(previous_items)
+
+        full_path_items.append(f'({previous_items_str})')
+    full_path_items.append(f'/{node_label}')
+    full_path = ''.join(full_path_items)
+    return full_path
